expose testing helper to other apps
